Add files via upload
An attempt at a gekko model with an objective. I think our objective could minimize the time it takes to land without going over acceleration bounds. A minimization in time would be like the jennings objective but we may instead want to minimize left over fuel or some other objective in addition to a proper landing. I haven't got this to work and have not yet included gimbal or angle, just x and y position and thrust.